+++ projects/pastanaga-angular/src/lib/table/table.module.ts
@@ -8,9 +8,13 @@ import { TableRowComponent } from './table-row/table-row.component';
 import { TableRowHeaderComponent } from './table-row-header/table-row-header.component';
 import { TableCellMenuComponent } from './table-cell-menu/table-cell-menu.component';
 import { PaFocusableModule } from '../focusable/focusable.module';
+import { TableCellHeaderSortableComponent } from './table-cell-header-sortable/table-cell-header-sortable.component';
+import { PaTranslateModule } from '../translate/translate.module';
+import { PaIconModule } from '../icon/icon.module';
+import { TableHeaderButtonComponent } from './table-cell-header-sortable/table-header-button.component';
 
 @NgModule({
-    imports: [CommonModule, PaFocusableModule],
+    imports: [CommonModule, PaFocusableModule, PaTranslateModule, PaIconModule],
     exports: [
         TableComponent,
         TableRowComponent,
@@ -18,6 +22,8 @@ import { PaFocusableModule } from '../focusable/focusable.module';
         TableRowHeaderComponent,
         TableHeaderDirective,
         TableCellMenuComponent,
+        TableCellHeaderSortableComponent,
+        TableHeaderButtonComponent,
     ],
     declarations: [
         TableComponent,
@@ -26,6 +32,8 @@ import { PaFocusableModule } from '../focusable/focusable.module';
         TableRowHeaderComponent,
         TableHeaderDirective,
         TableCellMenuComponent,
+        TableCellHeaderSortableComponent,
+        TableHeaderButtonComponent,
     ],
     providers: [],
 })