+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 2.17.9 (2021-01-11)
+
+### Improvement
+
+- **Table**: 
+    - Added sortable header component [niglesias]
+    - Added table header button to be used only in tables [niglesias]
 # 2.17.8 (2021-01-11)
 
 ### Improvements